Split channel state into subdirectory
